updated login and signup
